pyproject.toml
[project]
name = "downtime-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Cristian Ticu",email = "cristian.ticu@eggnita.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "boto3[crt] (>=1.37.15,<2.0.0)",
    "types-boto3[essential] (>=1.37.15,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
